src/requirements.txt
etext==0.0.1
Flask==1.1.2
python_amazon_sp_api==0.18.8
pytz==2022.1
